// client/src/type.ts
export type objectId = {
  _id: string;
};
export type User = {
  name: string;
  email: string;
  password: string;
  image: string;
};
export const facebookId = "426621496028034";
export const clientId =
  "1009282809407-sh8h2kgmot2q295a503sl5530pldnaj9.apps.googleusercontent.com";
//"491004959702-3bgqo54pt777f77dgl7cqd6s7e7rii81.apps.googleusercontent.com"